+++ src/packages/core/culture/components/index.ts
