+++ src/packages/core/culture/components/input-culture-select/input-culture-select.element.ts
