+++ src/packages/core/index.ts
@@ -8,7 +8,7 @@ import { manifests as workspaceManifests } from './workspace/manifests.js';
 import { manifests as modalManifests } from './modal/common/manifests.js';
 import { manifests as themeManifests } from './themes/manifests.js';
 import { manifests as conditionManifests } from './extension-registry/conditions/manifests.js';
-import { manifests as cultureManifests } from './cultures/manifests.js';
+import { manifests as cultureManifests } from './culture/manifests.js';
 
 import { UmbNotificationContext, UMB_NOTIFICATION_CONTEXT_TOKEN } from '@umbraco-cms/backoffice/notification';
 import { UmbModalManagerContext, UMB_MODAL_MANAGER_CONTEXT_TOKEN } from '@umbraco-cms/backoffice/modal';
